Correct the permission check
If I understand correctly, the permission that is needed is for $user/.acme.sh/$domain to be able to be copied to $certs.  This pull request makes that change in the script.

At any rate `/bin/cp -rf $user/.acme.sh/ .` should be avoided since the result for  will be different depending on which directory the `zimbra` user runs `/opt/letsencrypt/deploy-zimbra-letsencrypt.sh` from.